Refactor codebase architecture for maintainability and compilation speed
Decompose monolithic files, decouple subsystem boundaries, remove dead code, and optimize header dependencies across Illumo and IllumoGame:

- Phase 1 (Hygiene & Dead Code):
  * Remove dead unreferenced headers: CommonRenderTypes.h, IBuffer.h, VertexAttributeDescriptor.h, and Brush.h.
  * Delete obsolete RunCoverage.cmake and redundant repo-root editor-ui-atlas.jpg.
  * Harden TestIlscCodec and TestEditorModule with RAII TempFileGuard and safe atlas seeding to prevent file leaks.
  * Add missing headers (PlatformTimer.h, IModuleHost.h, RenderPass.h, RenderTargetPool.h, ShaderPreprocessor.h) to PublicHeaderSmoke.cpp.
  * Fix naming consistency for BriansBrainRuleSet.h with backwards-compatible alias.

- Phase 2 (Compilation Speed & Header Decoupling):
  * Create Renderer.cpp and extract non-template methods from Renderer.h, trimming preprocessor overhead across 35+ translation units.
  * Create CellContext.cpp, forward-declaring types in CellContext.h to eliminate transitive AllSets.h inclusion.
  * Decouple RuleSet.h from CellGrid.h via forward declaration, moving implementation to RuleSet.cpp.
  * Relocate PresentationTiming.h to public engine header Illumo/Include/Illumo/Engine/PresentationTiming.h with forwarding shim.

- Phase 3 (Monolith Decomposition & Boundary Alignment):
  * Extract multi-threaded worker pool from SparseCellGrid.cpp into SparseWorkerPool.h/.cpp.
  * Extract .illumo save/load persistence from CellGameModule.cpp into standalone IllumoCodec.h/.cpp.
  * Extract selection, cut/copy/paste, and transforms into CellClipboard.h/.cpp with 64-bit coordinate overflow guards.
  * Decouple command parsing and execution into CommandLineCore.h/.cpp, leaving CommandLine as a UI drawable.
  * Add dedicated unit tests for IllumoCodec direct roundtrip, clipboard edge cases, and mutual alias recursion depth.
