+++ Illumo/Source/Engine/DebugModule.cpp
@@ -1,7 +1,7 @@
 #ifndef GLFW_INCLUDE_NONE
 #define GLFW_INCLUDE_NONE
 #endif
-#include "Rendering/PresentationTiming.h"
+#include <Illumo/Engine/PresentationTiming.h>
 #include <GLFW/glfw3.h>
 #include <Illumo/Engine/DebugModule.h>
 #include <Illumo/Services/InputManager.h>
