--- Illumo/Source/Rendering/IBuffer.h
@@ -1,8 +1,0 @@
-#pragma once
-
-class IBuffer
-{
-public:
-  virtual ~IBuffer() = default;
-  virtual void SetData(const void* data, size_t size) = 0;
-};
