+++ Illumo/Source/Rendering/RenderWindow.cpp
@@ -1,5 +1,5 @@
 #include "RenderWindow.h"
-#include "PresentationTiming.h"
+#include <Illumo/Engine/PresentationTiming.h>
 #include <GL/glew.h>
 #include <GLFW/glfw3.h>
 #include <Illumo/Services/Logger.h>
