+++ Illumo/Tests/TestRuntimeUtilities.cpp
@@ -1,5 +1,5 @@
 #include "Rendering/BackendConfig.h"
-#include "Rendering/PresentationTiming.h"
+#include <Illumo/Engine/PresentationTiming.h>
 #include <GLFW/glfw3.h>
 #include <Illumo/Engine/IllumoContext.h>
 #include <Illumo/Rendering/AssetManager.h>
