+++ IllumoGame/Source/Game/SparseCellGrid.cpp
@@ -1,4 +1,5 @@
 #include "SparseCellGrid.h"
+#include "SparseWorkerPool.h"
 #include "Rulesets/RuleSet.h"
 #include <algorithm>
 #include <atomic>
@@ -36,18 +37,6 @@ millisecondsSince(const std::chrono::steady_clock::time_point& start)
            std::chrono::steady_clock::now() - start)
     .count();
 }
-
-struct SparseCellGrid::TargetResult
-{
-  ChunkAddress address;
-  CellArray cells{};
-  OccupancyMask occupied{};
-  OccupancyMask counted{};
-  OccupancyMask stateChanged{};
-  OccupancyMask countedChanged{};
-  bool hasNonBackground = false;
-  bool completeCells = false;
-};
 
 struct SparseCellGrid::ChunkMemoState
 {
@@ -202,277 +191,6 @@ SparseCellGrid::getCompleteResultCapacityForTesting() const
 {
   return m_completeResults.capacity();
 }
-
-class SparseWorkerPool
-{
-public:
-  SparseWorkerPool() = default;
-
-  ~SparseWorkerPool()
-  {
-    {
-      std::lock_guard<std::mutex> lock(mutex);
-      stopping = true;
-      workGeneration += 1;
-    }
-    workReady.notify_all();
-    for (std::thread& worker : workers) {
-      if (worker.joinable()) {
-        worker.join();
-      }
-    }
-  }
-
-  void evaluate(const SparseCellGrid* grid,
-                const unsigned char* transitions,
-                const std::vector<ChunkAddress>* targets,
-                std::vector<SparseCellGrid::TargetResult>* results,
-                unsigned int workerCount)
-  {
-    if (grid == nullptr || transitions == nullptr || targets == nullptr ||
-        results == nullptr || workerCount <= 1u || targets->empty()) {
-      return;
-    }
-
-    const unsigned int workerThreads = workerCount - 1u;
-    ensureWorkerCount(workerThreads);
-
-    {
-      std::lock_guard<std::mutex> lock(mutex);
-      activeGrid = grid;
-      activeTransitions = transitions;
-      activeTargets = targets;
-      activeResults = results;
-      activeCandidatePreparationScratch = nullptr;
-      activeCandidatePreparationRanges = nullptr;
-      activeCandidateScratch = nullptr;
-      activeCandidateRanges = nullptr;
-      nextWorkItem.store(0u);
-      availableWorkerSlots.store(workerThreads);
-      completedWorkers = 0u;
-      requiredWorkers = workerThreads;
-      workGeneration += 1;
-    }
-    workReady.notify_all();
-
-    executeAvailableWork(0u);
-
-    std::unique_lock<std::mutex> lock(mutex);
-    workComplete.wait(lock,
-                      [this]() { return completedWorkers >= requiredWorkers; });
-  }
-
-  void evaluateCandidates(
-    const SparseCellGrid* grid,
-    const unsigned char* transitions,
-    const std::vector<SparseCellGrid::CandidateScratchChunk>* scratch,
-    std::vector<SparseCellGrid::CandidateWorkRange>* ranges,
-    std::vector<SparseCellGrid::TargetResult>* results,
-    unsigned int workerCount)
-  {
-    if (grid == nullptr || transitions == nullptr || scratch == nullptr ||
-        ranges == nullptr || results == nullptr || workerCount <= 1u ||
-        ranges->empty()) {
-      return;
-    }
-
-    const unsigned int workerThreads = workerCount - 1u;
-    ensureWorkerCount(workerThreads);
-
-    {
-      std::lock_guard<std::mutex> lock(mutex);
-      activeGrid = grid;
-      activeTransitions = transitions;
-      activeTargets = nullptr;
-      activeCandidatePreparationScratch = nullptr;
-      activeCandidatePreparationRanges = nullptr;
-      activeCandidateScratch = scratch;
-      activeCandidateRanges = ranges;
-      activeResults = results;
-      nextWorkItem.store(0u);
-      availableWorkerSlots.store(workerThreads);
-      completedWorkers = 0u;
-      requiredWorkers = workerThreads;
-      workGeneration += 1;
-    }
-    workReady.notify_all();
-
-    executeAvailableWork(0u);
-
-    std::unique_lock<std::mutex> lock(mutex);
-    workComplete.wait(lock,
-                      [this]() { return completedWorkers >= requiredWorkers; });
-  }
-
-  void prepareCandidates(
-    const SparseCellGrid* grid,
-    std::vector<SparseCellGrid::CandidateScratchChunk>* scratch,
-    const std::vector<SparseCellGrid::CandidateWorkRange>* ranges,
-    unsigned int workerCount)
-  {
-    if (grid == nullptr || scratch == nullptr || workerCount <= 1u ||
-        ranges == nullptr || ranges->empty()) {
-      return;
-    }
-
-    const unsigned int workerThreads = workerCount - 1u;
-    ensureWorkerCount(workerThreads);
-
-    {
-      std::lock_guard<std::mutex> lock(mutex);
-      activeGrid = grid;
-      activeTransitions = nullptr;
-      activeTargets = nullptr;
-      activeCandidatePreparationScratch = scratch;
-      activeCandidatePreparationRanges = ranges;
-      activeCandidateScratch = nullptr;
-      activeCandidateRanges = nullptr;
-      activeResults = nullptr;
-      nextWorkItem.store(0u);
-      availableWorkerSlots.store(workerThreads);
-      completedWorkers = 0u;
-      requiredWorkers = workerThreads;
-      workGeneration += 1;
-    }
-    workReady.notify_all();
-
-    executeAvailableWork(0u);
-
-    std::unique_lock<std::mutex> lock(mutex);
-    workComplete.wait(lock,
-                      [this]() { return completedWorkers >= requiredWorkers; });
-  }
-
-private:
-  std::mutex mutex;
-  std::condition_variable workReady;
-  std::condition_variable workComplete;
-  std::vector<std::thread> workers;
-  std::atomic<std::size_t> nextWorkItem{ 0u };
-  std::atomic<unsigned int> availableWorkerSlots{ 0u };
-  const SparseCellGrid* activeGrid = nullptr;
-  const unsigned char* activeTransitions = nullptr;
-  const std::vector<ChunkAddress>* activeTargets = nullptr;
-  std::vector<SparseCellGrid::CandidateScratchChunk>*
-    activeCandidatePreparationScratch = nullptr;
-  const std::vector<SparseCellGrid::CandidateWorkRange>*
-    activeCandidatePreparationRanges = nullptr;
-  const std::vector<SparseCellGrid::CandidateScratchChunk>*
-    activeCandidateScratch = nullptr;
-  std::vector<SparseCellGrid::CandidateWorkRange>* activeCandidateRanges =
-    nullptr;
-  std::vector<SparseCellGrid::TargetResult>* activeResults = nullptr;
-  std::size_t workGeneration = 0u;
-  unsigned int requiredWorkers = 0u;
-  unsigned int completedWorkers = 0u;
-  bool stopping = false;
-
-  void ensureWorkerCount(unsigned int requiredCount)
-  {
-    while (workers.size() < static_cast<std::size_t>(requiredCount)) {
-      const unsigned int shardIndex =
-        static_cast<unsigned int>(workers.size()) + 1u;
-      workers.emplace_back(&SparseWorkerPool::workerLoop, this, shardIndex);
-    }
-  }
-
-  bool claimWorkerSlot()
-  {
-    unsigned int remaining = availableWorkerSlots.load();
-    while (remaining > 0u) {
-      if (availableWorkerSlots.compare_exchange_weak(remaining,
-                                                     remaining - 1u)) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  void executeAvailableWork(unsigned int memoShardIndex)
-  {
-    for (;;) {
-      const std::size_t index = nextWorkItem.fetch_add(1u);
-      if (activeGrid == nullptr) {
-        return;
-      }
-      if (activeCandidatePreparationScratch != nullptr) {
-        if (activeCandidatePreparationRanges == nullptr ||
-            index >= activeCandidatePreparationRanges->size()) {
-          return;
-        }
-        const SparseCellGrid::CandidateWorkRange& range =
-          (*activeCandidatePreparationRanges)[index];
-        for (std::size_t scratchIndex = range.begin; scratchIndex < range.end;
-             ++scratchIndex) {
-          activeGrid->prepareCandidateScratchChunk(
-            &(*activeCandidatePreparationScratch)[scratchIndex]);
-        }
-        continue;
-      }
-      if (activeResults == nullptr || activeTransitions == nullptr) {
-        return;
-      }
-      if (activeCandidateScratch != nullptr &&
-          activeCandidateRanges != nullptr) {
-        if (index >= activeCandidateRanges->size()) {
-          return;
-        }
-        SparseCellGrid::CandidateWorkRange& range =
-          (*activeCandidateRanges)[index];
-        range.outputChunkCount = 0u;
-        for (std::size_t scratchIndex = range.begin; scratchIndex < range.end;
-             ++scratchIndex) {
-          activeGrid->evaluateCandidateChunk(
-            (*activeCandidateScratch)[scratchIndex],
-            activeTransitions,
-            &(*activeResults)[scratchIndex],
-            memoShardIndex);
-          if ((*activeResults)[scratchIndex].hasNonBackground) {
-            range.outputChunkCount += 1u;
-          }
-        }
-        continue;
-      }
-      if (activeTargets == nullptr || index >= activeTargets->size()) {
-        return;
-      }
-      activeGrid->evaluateTargetChunk((*activeTargets)[index],
-                                      activeTransitions,
-                                      &(*activeResults)[index],
-                                      memoShardIndex);
-    }
-  }
-
-  void workerLoop(unsigned int memoShardIndex)
-  {
-    std::size_t observedGeneration = 0u;
-    for (;;) {
-      {
-        std::unique_lock<std::mutex> lock(mutex);
-        workReady.wait(lock, [this, &observedGeneration]() {
-          return stopping || workGeneration != observedGeneration;
-        });
-        if (stopping) {
-          return;
-        }
-        observedGeneration = workGeneration;
-      }
-
-      if (!claimWorkerSlot()) {
-        continue;
-      }
-      executeAvailableWork(memoShardIndex);
-
-      {
-        std::lock_guard<std::mutex> lock(mutex);
-        completedWorkers += 1u;
-        if (completedWorkers >= requiredWorkers) {
-          workComplete.notify_one();
-        }
-      }
-    }
-  }
-};
 
 SparseCellGrid::SparseCellGrid() = default;
